Anomalies class

+++ code/tests/robot/control/functions.ino
@@ -20,9 +20,9 @@ void receive_msg_pid() {
         messenger->parseMessage();
         float kp, kd, ki;
         int selection = (int) messenger->getMessage()[0];
-        kp = (int) messenger->getMessage()[1];
-        kd = (int) messenger->getMessage()[2];
-        ki = (int) messenger->getMessage()[3];
+        kp = (float) messenger->getMessage()[1];
+        kd = (float) messenger->getMessage()[2];
+        ki = (float) messenger->getMessage()[3];
 
         switch (int(selection)) {
             case 0: pid_speed->setParameters(kp, kd, ki); break;
@@ -40,9 +40,9 @@ void receive_msg_line() {
         messenger->parseMessage();
         float kp, kd, ki; 
         instruction = (int) messenger->getMessage()[0];
-        kp = (int) messenger->getMessage()[1];
-        kd = (int) messenger->getMessage()[2];
-        ki = (int) messenger->getMessage()[3];
+        kp = (float) messenger->getMessage()[1];
+        kd = (float) messenger->getMessage()[2];
+        ki = (float) messenger->getMessage()[3];
         pid_line->setParameters(kp, kd, ki); 
     }
 }
